+++ app/components/summaryOverview/__test__/SummaryOverviewBox.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
 import { describe, it, vi, expect } from "vitest";
 import { type HeadingProps } from "~/components/Heading";
-import { useFormFlow } from "../../form/formFlowContext";
+import { useFormFlow } from "../../formElements/formFlowContext";
 import SummaryOverviewBox from "../SummaryOverviewBox";
 
 vi.mock("../../form/formFlowContext", () => ({
