+++ app/components/summaryOverview/__test__/SummaryOverviewBoxWrapped.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import { useFormFlow } from "../../form/formFlowContext";
+import { useFormFlow } from "../../formElements/formFlowContext";
 import SummaryOverviewBoxWrapped from "../SummaryOverviewBoxWrapped";
 
 vi.mock("../../form/formFlowContext", () => ({
