+++ app/routes/shared/components/FormFlowPage.tsx
@@ -3,8 +3,8 @@ import { useLoaderData } from "react-router";
 import ArraySummary from "~/components/arraySummary/ArraySummary";
 import Background from "~/components/Background";
 import ContentComponents from "~/components/ContentComponents";
-import { FormFlowContext } from "~/components/form/formFlowContext";
-import ValidatedFlowForm from "~/components/form/ValidatedFlowForm";
+import { FormFlowContext } from "~/components/formElements/formFlowContext";
+import ValidatedFlowForm from "~/components/formElements/ValidatedFlowForm";
 import Heading from "~/components/Heading";
 import { useFocusFirstH1 } from "~/components/hooks/useFocusFirstH1";
 import MigrationDataOverview from "~/components/MigrationDataOverview";
