+++ app/routes/shared/components/VorabcheckPage.tsx
@@ -1,7 +1,7 @@
 import { useLoaderData } from "react-router";
 import Background from "~/components/Background";
 import ContentComponents from "~/components/ContentComponents";
-import ValidatedFlowForm from "~/components/form/ValidatedFlowForm";
+import ValidatedFlowForm from "~/components/formElements/ValidatedFlowForm";
 import { useFocusFirstH1 } from "~/components/hooks/useFocusFirstH1";
 import { ProgressBar } from "~/components/ProgressBar";
 import { ReportProblem } from "~/components/reportProblem/ReportProblem";
